[refactor] : #134 버튼 클릭 수정 및 불필요 파일 삭제
