--- feature/clip/src/main/java/org/sopt/clip/ClipsDTO.kt
@@ -1,7 +1,0 @@
-package org.sopt.clip
-
-data class ClipsDTO(
-  val clipName: String,
-  val clipAmount: Int,
-  val clipId: Long,
-)
