--- feature/clip/src/main/java/org/sopt/clip/LinkDTO.kt
@@ -1,9 +1,0 @@
-package org.sopt.clip
-
-data class LinkDTO(
-  val url: String,
-  val thumbnail: Int,
-  val linkTitle: String,
-  val upperClipName: String,
-  val linkId: Long,
-)
