+++ feature/savelink/src/main/java/org/sopt/savelink/ui/savelink/SaveLinkFragment.kt
@@ -175,6 +175,7 @@ class SaveLinkFragment : BindingFragment<FragmentSaveLinkBinding>({ FragmentSave
 
   private fun onClickComplete() {
     binding.btnSaveLinkNext.btnClick {
+      if (binding.btnSaveLinkNext.state!=LinkMIndFullWidthButtonState.DISABLE)
       viewModel.navigateSetLink()
     }
   }
